factor out conntrack GC into separate package
CT garbage collection uses the EndpointManager API - it does not need to be
tied into `pkg/endpointmanager`. Remove it accordingly.

Signed-off by: Ian Vernon <ian@cilium.io>
